chore(mps-sync-plugin): move classes to separate files
